chore: Rename defs files
