--- src/lib/app/defs.ts
@@ -1,1 +1,0 @@
-export * from '../commonDefs'
