+++ src/lib/index.app.ts
@@ -2,5 +2,5 @@
 
 export * from './app/useQueryState'
 export * from './app/useQueryStates'
-export * from './commonDefs'
+export * from './defs'
 export const __useQueryStateRouter = 'app'
